requirements.txt
python-telegram-bot[job-queue]>=21.0
python-dotenv
requests
web3
py-clob-client
eth-account
rich
httpx
apscheduler
